feat: refactor schema rules
