+++ src/schema/permissions/index.ts
@@ -1,27 +1,5 @@
-import { allow, rule, shield } from 'graphql-shield'
-import { Context, getUserId } from '../../utils'
-
-const rules = {
-  isAuthenticatedUser: rule()(
-    (_parent: unknown, _args: unknown, context: Context) => {
-      const userId = getUserId(context)
-      return Boolean(userId)
-    },
-  ),
-  isPostOwner: rule()(
-    async (_parent: unknown, { id }: { id: number }, context) => {
-      const userId = getUserId(context)
-      const author = await context.prisma.post
-        .findOne({
-          where: {
-            id: Number(id),
-          },
-        })
-        .author()
-      return userId === author.id
-    },
-  ),
-}
+import { allow, shield } from 'graphql-shield'
+import rules from './rules'
 
 export const permissions = shield({
   Query: {
